Add a minimum walking multipath iterator
Create an iterator that when given multiple paths will walk the minimum
tree distance to return a reference to each node at that path. The
iterator utilizes a stack and does a single allocation for the stack at
the beginning of operation to the depth of the tree so that it doesn't
require resizing.

Also create a mutable find operation that will return a mutable
reference to a node along with the a new tree that has a copied path to
get to the mutable node. This allows for abstracting the tree walking
and creation for single node update operations.