WIP - Add snapshot func and helpers for restore
This commit is for safekeeping. The snapshot function is implemented and
will write out a tree in lexicographic order with each node msgpack
formatted. The helper functions that allow reading and constructing a
Node from the msgpack data are also in this commit. However, the
function that takes these nodes and reconstructs the tree is not yet
written. That, and property based tests for snapshot and restore round
trips remain for the functionality to be considered complete.