Add multi-cas support and a bunch of other stuff
Fix build: Remove serde and quickcheck_macros
Add Delete operation and verify path for all ops

Remove all uses of RefCell

RefCell was used to provide "interior mutability" to Nodes, so that
durin COW operations such as insert, delete, etc.. the newly created
nodes could be mutated such that children could be linked to their
parents. This allows maintaining logical immutability of the tree while
still allowing mutation during construction.

However, using RefCell makes the tree `!Sync`, meaning that it can't be
transferred across threads. This is unnacceptable since the tree is
logically immutable and one of the main purposes of a COW tree is that
it can be safely shared across threads for reads.

The `!Sync` issue could be fixed by explicitly declaring the tree `Sync`
and overriding the compiler. It's unclear to me what consequences this
would have, although it's likely safe.

Alternatively, RefCell could just be removed altogether and replaced
with unsafe code during tree construction. Due to the fact that RefCell
was not enough to replace all uses of unsafe code (specifically during
mutable iteration), and it already was somewhat prevelant in the
codebase, this was the solution I chose. This also has the side effect
of removing all the obfuscating `.borrow()` calls in the same functions
containing unsafe code.

Remove extra path validations, fix bug in tree.find

Add a property based test for tree.find and PathIter

Fix PathIter and set operations

Ensure PathIter can iterate over nodes at the same level in the tree when the
path of one node is a prefix of the path of another.

Set Intersection, difference, and symmetric difference aren't actually
commutative. Change the signature to only accept to paths.

Add `Empty` Value