Add rustfmt.toml and run rustfmt against code
The config is not perfectly matched to my personal style, as rustfmt is
missing options allowing full configurability. There are still trailing
commas on match blocks where the resulting expression is bound to a
variable using `let` and it doesn't seem possible to get rid of them.
There are other places, like a space between braces in struct literals
and the first character, that also don't seem configurable. However, I'm
willing to live with these slight style disagreements in order to have
the benefit of a machine checking.

Here is a proposed configuration for rustfmt and the resulting output.
The escape hatch to prevent mangling:
`#[cfg_attr(rustfmt, rustfmt_skip)]` was only used once. in an effort to
visually compress a match statement.